requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
cffi==1.14.2
Django==3.1
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-timezone-field==4.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
kombu==4.6.11
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
rabbitmq==0.2.0
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
